Don't log deprecated notices
There is a breaking behavioural change in PHP 5.6 that will result in getting valid requests logged as DEPRECATED notice which in some cases can result in broken AJAX requests and other unpredictable havoc.

As a workaround we can change the error_reporting handler to exclude deprecated notices. The actual fix would be to adjust the ini setting but since this is not possible programmatically due to the fact that user code is executed after this exception is thrown we have to do it this way. Our .htaccess and .user.ini will still contain the correct fix.

Fixes https://github.com/owncloud/core/issues/16014